+++ R/build.R
@@ -10,6 +10,7 @@
 #' @param meta list of metadata that tell which
 #'   targets are up to date (from [drake_meta()]).
 #' @param config internal configuration list
+#' @inheritParams loadd
 #' @examples
 #' \dontrun{
 #' test_with_dir("Quarantine side effects.", {
@@ -19,22 +20,54 @@
 #' load_basic_example() # Get the code with drake_example("basic").
 #' # Create the master internal configuration list.
 #' config <- drake_config(my_plan)
-#' # Optionally, compute metadata on 'small',
-#' # including a hash/fingerprint
-#' # of the dependencies. If meta is not supplied,
-#' # drake_build() computes it automatically.
-#' meta <- drake_meta(target = "small", config = config)
-#' # Should not yet include 'small'.
+#' out <- drake_build(small, config = config)
+#' # Now includes `small`.
 #' cached()
-#' # Build 'small'.
-#' # Equivalent to just drake_build(target = "small", config = config).
-#' drake_build(target = "small", config = config, meta = meta)
-#' # Should now include 'small'
-#' cached()
-#' readd(small)
+#' head(readd(small))
+#' # `small` was invisibly returned.
+#' head(out)
+#' # If you previously called make(),
+#' # `config` is just read from the cache.
+#' make(my_plan, verbose = FALSE)
+#' result <- drake_build(small)
+#' head(result)
 #' })
 #' }
-drake_build <- function(target, config, meta = NULL){
+drake_build <- function(
+  target,
+  config = NULL,
+  meta = NULL,
+  character_only = FALSE,
+  envir = parent.frame(),
+  jobs = 1,
+  replace = FALSE
+){
+  if (!is.null(meta)){
+    warning(
+      "drake_build() is exclusively user-side now, ",
+      "so we can affort to compute `meta` on the fly. ",
+      "Thus, the `meta` argument is deprecated."
+    )
+  }
+  if (!character_only){
+    target <- as.character(substitute(target))
+  }
+  if (is.null(config)){
+    config <- read_drake_config(envir = envir, jobs = jobs)
+  }
+  loadd(
+    list = target,
+    deps = TRUE,
+    envir = envir,
+    cache = config$cache,
+    graph = config$graph,
+    jobs = jobs,
+    replace = replace
+  )
+  build_and_store(target = target, config = config)
+}
+
+build_and_store <- function(target, config, meta = NULL){
   # The environment should have been pruned by now.
   # For staged parallelism, this was already done in bulk
   # for the whole stage.
@@ -45,19 +78,7 @@ drake_build <- function(target, config, meta = NULL){
       meta <- drake_meta(target = target, config = config)
     }
     announce_build(target = target, meta = meta, config = config)
-    if (meta$imported) {
-      value <- process_import(target = target, config = config)
-    } else {
-      # build_target() does not require access to the cache.
-      # A custom future-based job scheduler could build with different steps
-      # to write the output to the master process before caching it.
-      value <- build_target(
-        target = target,
-        meta = meta,
-        start = start,
-        config = config
-      )
-    }
+    value <- just_build(target = target, meta = meta, config = config)
     conclude_build(
       target = target,
       value = value,
@@ -66,6 +87,21 @@ drake_build <- function(target, config, meta = NULL){
       config = config
     )
   })
+}
+
+just_build <- function(target, meta, config){
+  if (meta$imported) {
+    process_import(target = target, config = config)
+  } else {
+    # build_target() does not require access to the cache.
+    # A custom future-based job scheduler could build with different steps
+    # to write the output to the master process before caching it.
+    build_target(
+      target = target,
+      meta = meta,
+      config = config
+    )
+  }
 }
 
 announce_build <- function(target, meta, config){
@@ -90,7 +126,7 @@ conclude_build <- function(target, value, meta, start, config){
   invisible(value)
 }
 
-build_target <- function(target, meta, start, config) {
+build_target <- function(target, meta, config) {
   command <- get_evaluation_command(target = target, config = config)
   seed <- list(seed = config$seed, target = target) %>%
     seed_from_object